requirements.txt
aiogram==3.15.0
APScheduler==3.10.4
python-dotenv==1.0.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
